requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap4==2023.1
cycler==0.12.1
dj-database-url==2.3.0
Django==4.2.7
django-crispy-forms==2.1
fonttools==4.49.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
numpy==1.26.1
packaging==24.0
pandas==2.1.3
pillow==10.2.0
psycopg2-binary==2.9.10
pyparsing==3.1.2
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sk==0.0.1
sqlparse==0.4.4
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
uvicorn==0.34.0
whitenoise==6.9.0
